requirements.txt
emails
google-api-python-client
google-auth
mail-parser
maybe-else
pathmagic
pysubtypes
pymiscutils
pyiotools
